+++ src/main/java/net/sf/jabref/gui/FieldContentSelector.java
@@ -41,7 +41,7 @@ import net.sf.jabref.logic.l10n.Localization;
 /**
  * A combo-box and a manage button that will add selected strings to an
  * associated entry editor.
- * 
+ *
  * Used to manage keywords and authors for instance.
  */
 public class FieldContentSelector extends JComponent {
@@ -61,9 +61,9 @@ public class FieldContentSelector extends JComponent {
 
 
     /**
-     * 
+     *
      * Create a new FieldContentSelector.
-     * 
+     *
      * @param frame
      *            The one JabRef-Frame.
      * @param panel
@@ -137,7 +137,7 @@ public class FieldContentSelector extends JComponent {
                  * list, so we should not react to it. I'm not sure if this is
                  * well defined enough to be guaranteed to work everywhere.
                  */
-                if ("comboBoxChanged".equals(e.getActionCommand()) && e.getModifiers() == 0) {
+                if ("comboBoxChanged".equals(e.getActionCommand()) && (e.getModifiers() == 0)) {
                     return;
                 }
 
@@ -190,7 +190,7 @@ public class FieldContentSelector extends JComponent {
         }
 
         String chosen = (String) comboBox.getSelectedItem();
-        if (chosen == null || "".equals(chosen)) {
+        if ((chosen == null) || chosen.isEmpty()) {
             return;
         }
 
@@ -236,7 +236,7 @@ public class FieldContentSelector extends JComponent {
     //	/**
     //	 * Adds a word to the selector (to the JList and to the MetaData), unless it
     //	 * is already there
-    //	 * 
+    //	 *
     //	 * @param newWord
     //	 *            String Word to add
     //	 */