+++ src/main/java/net/sf/jabref/gui/preftabs/AppearancePrefsTab.java
@@ -20,6 +20,7 @@ import java.awt.Font;
 import java.awt.GridBagLayout;
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
+import java.util.Optional;
 
 import javax.swing.*;
 
@@ -120,11 +121,8 @@ class AppearancePrefsTab extends JPanel implements PrefsTab {
 
             @Override
             public void actionPerformed(ActionEvent e) {
-                Font f = new FontSelectorDialog
-                        (null, GUIGlobals.CURRENTFONT).getSelectedFont();
-                if (f != null) {
-                    font = f;
-                }
+                Optional<Font> f = new FontSelectorDialog(null, GUIGlobals.CURRENTFONT).getSelectedFont();
+                f.ifPresent(x -> font = x);
             }
         });