+++ src/test/java/net/sf/jabref/model/entry/EntryUtilTest.java
@@ -1,6 +1,6 @@
 package net.sf.jabref.model.entry;
 
-import java.util.ArrayList;
+import java.util.List;
 
 import org.junit.Assert;
 import org.junit.Test;
@@ -18,7 +18,7 @@ public class EntryUtilTest {
     @Test
     public void getSeparatedKeywords() {
         String keywords = "w1, w2a w2b, w3";
-        ArrayList<String> separatedKeywords = EntryUtil.getSeparatedKeywords(keywords);
+        List<String> separatedKeywords = EntryUtil.getSeparatedKeywords(keywords);
         String[] expected = new String[] {"w1", "w2a w2b", "w3"};
         Assert.assertArrayEquals(expected, separatedKeywords.toArray());
     }